+++ totalRP3/modules/dashboard/WhatsNewTabView.lua
@@ -39,13 +39,9 @@ local strhtml = TRP3_API.utils.str.toHTML;
 --- List of "What's new" segments to concatenate and display together when
 ---  this view is activated.
 local SEGMENTS = {
+	loc.WHATS_NEW_24_13,
 	loc.WHATS_NEW_24_12,
 	loc.WHATS_NEW_24_11,
-	loc.WHATS_NEW_24_10,
-	loc.WHATS_NEW_24_9,
-	loc.WHATS_NEW_24_8,
-	loc.WHATS_NEW_24_7,
-	loc.WHATS_NEW_24_6,
 };
 
 --- Returns the fully formatted localized text for this view.